Rename to RemoveTestsOverriddenPrivateMethodParameterRector, mark test scope in name
