+++ rules-tests/DeadCode/Rector/ClassMethod/RemoveTestsOverriddenPrivateMethodParameterRector/Fixture/overridden_param_with_php_doc.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveOverriddenPrivateMethodParameterRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveTestsOverriddenPrivateMethodParameterRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
 
@@ -26,7 +26,7 @@ final class OverriddenParamWithPhpDoc extends TestCase
 -----
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveOverriddenPrivateMethodParameterRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveTestsOverriddenPrivateMethodParameterRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
