+++ rules-tests/DeadCode/Rector/ClassMethod/RemoveTestsOverriddenPrivateMethodParameterRector/Fixture/skip_by_ref_override.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveOverriddenPrivateMethodParameterRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveTestsOverriddenPrivateMethodParameterRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
