+++ src/Config/Level/DeadCodeLevel.php
@@ -19,10 +19,10 @@ use Rector\DeadCode\Rector\ClassMethod\RemoveDuplicatedReturnSelfDocblockRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveEmptyClassMethodRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveMixedDocblockOverruledByNativeTypeRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveNullTagValueNodeRector;
-use Rector\DeadCode\Rector\ClassMethod\RemoveOverriddenPrivateMethodParameterRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveParentDelegatingClassMethodRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveParentDelegatingConstructorRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveReturnTagIncompatibleWithNativeTypeRector;
+use Rector\DeadCode\Rector\ClassMethod\RemoveTestsOverriddenPrivateMethodParameterRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveUnusedConstructorParamRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveUnusedPrivateMethodParameterRector;
 use Rector\DeadCode\Rector\ClassMethod\RemoveUnusedPrivateMethodRector;
@@ -165,7 +165,7 @@ final class DeadCodeLevel
         // removing methods could be risky if there is some magic loading them
         RemoveUnusedPromotedPropertyRector::class,
         RemoveUnusedPrivateMethodParameterRector::class,
-        RemoveOverriddenPrivateMethodParameterRector::class,
+        RemoveTestsOverriddenPrivateMethodParameterRector::class,
         RemoveUnusedPublicMethodParameterRector::class,
         RemoveUnusedPrivateMethodRector::class,
         RemoveUnreachableStatementRector::class,
